+++ internal/cel/transformer.go
@@ -1,4 +1,4 @@
-package transformer
+package cel
 
 import (
 	"errors"
