+++ internal/cel/transformer_test.go
@@ -1,4 +1,4 @@
-package transformer
+package cel
 
 import (
 	"testing"
